requirements.txt
alabaster==0.7.10

astroid==1.6.3

attrs==18.1.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1

Bottleneck==1.2.1
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
Django==2.0.8
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-filter==2.0.0
markdown==3.0.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
markdown2==2.3.7

contextlib2==0.5.5
cryptography==2.4.2
cycler==0.10.0
Cython==0.28.2

dask==0.17.5

decorator==4.3.0
distributed==1.21.8
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
geographiclib==1.49
geopy==1.14.0
gevent==1.3.0
glob2==0.6
greenlet==0.4.13
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.3.0
imagesize==1.0.0
ipykernel==4.8.2

itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0

kiwisolver==1.0.1
lazy-object-proxy==1.3.1

lxml==4.2.1
itsdangerous==0.24

Jinja2==2.10
MarkupSafe==1.0

matplotlib==2.2.2
mccabe==0.6.1

mistune==0.8.3

more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0

nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.14.3
numpydoc==0.8.0


opencv-python==3.4.1.15
openpyxl==2.5.3
packaging==17.1
pandas==0.23.0
pandas-datareader==0.6.0
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
prompt-toolkit==1.0.15
psutil==5.4.5
psycopg2==2.7.5
py==1.5.3
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18

psycopg2==2.7.5

pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4

pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2

pywinpty==0.5.1
PyYAML==4.2b1
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
rope==0.11.0
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4

xlwt==1.3.0
zict==0.1.3
Send2Trash==1.5.0